beifen

+++ diansai/MDK-ARM/BSP/BSP_H/bsp_usart.h
@@ -14,7 +14,7 @@ float get_yaw_atk901(void);
 float get_yaw_speed(void);
 float get_Acc_x(void);
 uint8_t*get_s_rx_buf_u8(void);
-uint8_t get_ball_x(void);
+uint16_t get_ball_x(void);
 
 #endif
