+++ diansai/MDK-ARM/BSP/bsp_usart.c
@@ -139,7 +139,7 @@ uint8_t*get_s_rx_buf_u8(void)
 {
   return s_rx_buf_u8;
 }
-uint8_t get_ball_x(void)
+uint16_t get_ball_x(void)
 {
   return ball_x;
 }
